Refactor authenticate
